+++ src/components/auth/SignUp/SignupForm.jsx
@@ -6,7 +6,7 @@ import { IconX, IconCheck } from '@tabler/icons-react';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { notifications } from '@mantine/notifications';
 
-import { signUpSchema } from '../../schema/schema';
+import { signUpSchema } from '../../../schema/schema';
 import Typing from './Typing';
 
 import SignupInput from './SignupInput';
