+++ src/components/auth/SignUp/Typing.jsx
