+++ src/components/auth/SignUp/index.js
@@ -1,0 +1,6 @@
+export { default as SignupHeader } from './SignupHeader';
+export { default as SignupOption } from './SignupOption';
+export { default as SignupForm } from './SignupForm';
+
+export { default as LogoBtn } from './LogoBtn';
+export { default as Typing } from './Typing';
