+++ src/components/auth/index.js
@@ -1,6 +1,5 @@
 export { default as SigninForm } from './SigninForm';
-export { default as SignupHeader } from './SignupHeader';
-export { default as SignupForm } from './SignupForm';
-export { default as SignupOption } from './SignupOption';
-export { default as Typing } from './Typing';
-export { default as LogoBtn } from './LogoBtn';
+
+export { default as SignupHeader } from './SignUp/SignupHeader';
+export { default as SignupOption } from './SignUp/SignupOption';
+export { default as SignupForm } from './SignUp/SignupForm';
